+++ PluginsAndFeatures/azure-toolkit-for-rider/azure-intellij-plugin-appservice-dotnet/src/main/kotlin/com/microsoft/azure/toolkit/intellij/legacy/webapp/runner/webappconfig/WebAppConfigurationType.kt
@@ -22,4 +22,6 @@ class WebAppConfigurationType : SimpleConfigurationType(
 
     override fun createConfiguration(name: String?, template: RunConfiguration) =
         WebAppConfiguration(template.project, this, name)
+
+    override fun getOptionsClass() = WebAppConfigurationOptions::class.java
 }